requirements.txt
tensorrt
onnx
onnxruntime 
pytorch 
cryptography 
numpy
torch2trt